fixed test

+++ src/main/scala/pl/mbadziong/flight/FlightRequest.scala
@@ -2,4 +2,4 @@ package pl.mbadziong.flight
 
 import pl.mbadziong.drone.Position
 
-case class FlightRequest(id: Int, route: List[Position])
+case class FlightRequest(id: Long, route: List[Position])
