+++ src/main/scala/pl/mbadziong/flight/FlightResponse.scala
@@ -1,3 +1,3 @@
 package pl.mbadziong.flight
 
-case class FlightResponse(id: Int)
+case class FlightResponse(id: Long)
